--- doc-cloud-samples.ui/src/main/webapp/extractpdf/index.html
@@ -1,30 +1,0 @@
-<html>
-    <head>
-        <title>Extract PDF</title>
-    </head>
-    <body>
-        <h1>Invoke Extract PDF Doc Cloud API</h1>
-
-        <form method="POST" action="/doc-cloud/samples/textextractpdf" enctype="multipart/form-data">
-            <label for="accountId">Account ID:</label>
-            <input type="text" id="accountId" name="accountId">
-
-            <label for="clientId">Client ID:</label>
-            <input type="text" id="clientId" name="clientId">
-
-            <label for="clientSecret">Client Secret:</label>
-            <input type="text" id="clientSecret" name="clientSecret">
-
-            <label for="organizationId">Organization ID:</label>
-            <input type="text" id="organizationId" name="organizationId">
-
-            <label for="key">Key:</label>
-            <input type="file" id="key" name="key">
-
-            <label for="pdf">PDF:</label>
-            <input type="file" id="pdf" name="pdf">
-
-            <input type="submit" value="Submit">
-        </form>
-    </body>
-</html>
