+++ module-template.ui/pom.xml
@@ -2,15 +2,15 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <artifactId>doc-cloud-samples</artifactId>
-    <groupId>com.aftia.application</groupId>
+    <groupId>com.aftia.vm.system</groupId>
+    <artifactId>module-template</artifactId>
     <version>1.0-SNAPSHOT</version>
   </parent>
 
-  <groupId>com.aftia.application</groupId>
-  <artifactId>doc-cloud-samples.ui</artifactId>
+  <groupId>com.aftia.vm.system</groupId>
+  <artifactId>module-template.ui</artifactId>
 
-  <name>Doc Cloud Samples - UI</name>
+  <name>Module Template - UI</name>
   
   <packaging>bundle</packaging>
