+++ module-template.ui/src/main/webapp/extractpdf/index.html
@@ -1,0 +1,15 @@
+<html>
+    <head>
+        <title>Module Template</title>
+    </head>
+    <body>
+        <h1>Module Template</h1>
+
+        <form method="POST" action="/samples/message" enctype="multipart/form-data">
+            <label for="message">Message:</label>
+            <input type="text" id="message" name="message">
+
+            <input type="submit" value="Submit">
+        </form>
+    </body>
+</html>
